requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
asyncpg==0.30.0
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
Flask==3.0.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.1
peewee==3.17.7
propcache==0.2.0
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pypika-tortoise==0.2.1
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
sniffio==1.3.1
typing_extensions==4.12.2
Werkzeug==3.1.1
yarl==1.17.1
